Abstracted in the Weaviate provider all the business logic. Now Oraculum only invoke the default provider.
