+++ .gitignore
@@ -369,3 +369,5 @@ FodyWeavers.xsd
 # Editors config files
 .idea/
 .vscode/
+/demo/web/docker-compose.secret.yml
+/demo/web/oraculumApiAppsettings.secret.json
